refactored app.jsx to use player.jsx logic
